Introduce callbacks to hide and show container.
Instead of using the "on" callbacks that for hide and show behavior,
allowing a callback will prevent the original functions to be called.

So for example, if I want to use slideUp/slideDown the animate effect will not be
performed because the show/hide will be performed before the animated functions.